requirements.txt
requests==2.31.0
roblox==2.0.0
rblx-open-cloud @ git+https://github.com/lawmixer/opencloud-rblx.git@v2
pytz==2023.3
Quart==0.19.4
nest-asyncio==1.5.8
firebase-admin==6.2.0
discord-py-interactions==5.11.0
aiofiles==23.2.1
aiohttp==3.9.2
aiosignal==1.3.1
asgiref==3.7.2
hypercorn==0.16.0
feedparser==6.0.10
uvicorn==0.25.0
redis==5.0.1
python-dotenv==1.0.1
dnspython>=2.6.1